web/api/deploy/app/requirements.txt
Flask==1.1.2
pyjwt==2.1.0